Flatten `red_knot_project` import paths
