+++ crates/red_knot/src/main.rs
@@ -6,11 +6,10 @@ use clap::Parser;
 use colored::Colorize;
 use crossbeam::channel as crossbeam_channel;
 use python_version::PythonVersion;
-use red_knot_project::db::ProjectDatabase;
-use red_knot_project::project::options::{EnvironmentOptions, Options};
-use red_knot_project::project::ProjectMetadata;
+use red_knot_project::metadata::options::{EnvironmentOptions, Options};
 use red_knot_project::watch;
 use red_knot_project::watch::ProjectWatcher;
+use red_knot_project::{ProjectDatabase, ProjectMetadata};
 use red_knot_python_semantic::SitePackages;
 use red_knot_server::run_server;
 use ruff_db::diagnostic::Diagnostic;
