+++ crates/red_knot_project/src/metadata/pyproject.rs
@@ -1,6 +1,60 @@
+use pep440_rs::{Version, VersionSpecifiers};
 use serde::{Deserialize, Deserializer, Serialize};
 use std::ops::Deref;
 use thiserror::Error;
+
+use crate::metadata::options::Options;
+
+/// A `pyproject.toml` as specified in PEP 517.
+#[derive(Deserialize, Serialize, Debug, Default, Clone)]
+#[serde(rename_all = "kebab-case")]
+pub struct PyProject {
+    /// PEP 621-compliant project metadata.
+    pub project: Option<Project>,
+    /// Tool-specific metadata.
+    pub tool: Option<Tool>,
+}
+
+impl PyProject {
+    pub(crate) fn knot(&self) -> Option<&Options> {
+        self.tool.as_ref().and_then(|tool| tool.knot.as_ref())
+    }
+}
+
+#[derive(Error, Debug)]
+pub enum PyProjectError {
+    #[error(transparent)]
+    TomlSyntax(#[from] toml::de::Error),
+}
+
+impl PyProject {
+    pub(crate) fn from_toml_str(content: &str) -> Result<Self, PyProjectError> {
+        toml::from_str(content).map_err(PyProjectError::TomlSyntax)
+    }
+}
+
+/// PEP 621 project metadata (`project`).
+///
+/// See <https://packaging.python.org/en/latest/specifications/pyproject-toml>.
+#[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
+#[serde(rename_all = "kebab-case")]
+pub struct Project {
+    /// The name of the project
+    ///
+    /// Note: Intentionally option to be more permissive during deserialization.
+    /// `PackageMetadata::from_pyproject` reports missing names.
+    pub name: Option<PackageName>,
+    /// The version of the project
+    pub version: Option<Version>,
+    /// The Python versions this project is compatible with.
+    pub requires_python: Option<VersionSpecifiers>,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
+#[serde(rename_all = "kebab-case")]
+pub struct Tool {
+    pub knot: Option<Options>,
+}
 
 /// The normalized name of a package.
 ///
