+++ crates/red_knot_project/tests/check.rs
@@ -1,6 +1,5 @@
 use anyhow::{anyhow, Context};
-use red_knot_project::db::ProjectDatabase;
-use red_knot_project::project::ProjectMetadata;
+use red_knot_project::{ProjectDatabase, ProjectMetadata};
 use red_knot_python_semantic::{HasTy, SemanticModel};
 use ruff_db::files::{system_path_to_file, File};
 use ruff_db::parsed::parsed_module;
